requirements.txt
streamlit
apify-client
langchain
langchain-openai
langchain-community
python-dotenv   
requests
beautifulsoup4
google-search-results
openai
chromadb
numpy
pandas
pydantic
tiktoken
fastapi
uvicorn
asyncio
dataclasses

